+++ R/aaa02_squarebrackets_methods.R
@@ -20,21 +20,12 @@
 #'  - `_x`: extract, exchange, or duplicate (if applicable) subsets.
 #'  - `_wo`: returns the original object \bold{without} the provided subsets.
 #'  - `_mod`: modify subsets and return copy.
-#'  - `_set`: modify subsets using \link[=squarebrackets_PassByReference]{pass-by-reference semantics}.
-#'  - `_rec`: access recursive subsets of lists - ONLY available with the `i2_` prefix.
-#'  - `_recin`: replace, transform, delete, or extend recursive subsets of lists - ONLY available with the `i2_` prefix. \cr \cr
+#'  - `_set`: modify subsets using \link[=squarebrackets_PassByReference]{pass-by-reference semantics}. \cr \cr
 #' 
-#' There is also the somewhat separate \link{idx} method,
-#' which works on both recursive and non-recursive objects. \cr
-#' 
-#' Thus, for example,
-#' the \link{i_x} method extracts/exchanges/duplicates subsets of atomic objects by flat/linear indices. \cr \cr
-#' 
-#' 
-#' @section Illustration: 
 #' 
 #' To illustrate, let's take the methods used for extracting subsets
-#' (`_x`): \cr
+#' (`_x`). \cr
+#' \cr
 #' When `y` is atomic, the following holds (roughly speaking):
 #' 
 #'  - `i_x(y, i)` corresponds to `y[i]`
@@ -46,17 +37,33 @@
 #'  - `ss2_x(y, n(i, k), c(1, 3))` corresponds to `y[i, , k]` or `y[[i, , k]]` (depending on the arguments given in `ss2_x()`) \cr \cr
 #' 
 #' 
+#' @section Other Methods:
+#' Besides the main methods, 'squarebrackets' provides some additional methods that do not neatly fit into the above methods. \cr
+#' \cr
+#' First, there is the \link[=lst_rec]{lst_} set of methods,
+#' which deal with sub-set operations that are only relevant for (nested) lists,
+#' but not for the other types of supported objects. \cr
+#' \cr
+#' Second, there is \link{idx} method,
+#' which works on both recursive and non-recursive objects,
+#' and transforms/translates indices to be used in R's default copy-on-modify semantics. \cr
+#' \cr
+#' Finally, there are the `sb_` and `sb2_` sets of methods,
+#' which cover miscellaneous operations for atomic and recursive objects,
+#' respectively. \cr \cr
+#' 
+#' 
 #' @section Finding the Appropriate Help Pages:
 #' 
 #' Now that it is clear what the naming convention is of the main methods,
 #' one can easily find out information about a particular method by usign the `?` operator. \cr
-#' So to find out about modifying recursive objects by reference using dimensional indices,
+#' So to find out about modifying recursive objects by subscripts using Pass-by-Reference semantics,
 #' type in: \cr
 #' `?ss2_set` \cr
 #' \cr
 #' Note that some combinations of `A` and `B` are not available. \cr
 #' For example, as the `_rec` `_recin` suffices are only applicable for recursive objects using linear indices,
-#' only `i2_rec` and `i2_recin` exist; \cr
+#' only `lst_rec` and `lst_recin` exist; \cr
 #' combinations like `i_rec`, `ss_recin`, etc. don't exist. \cr
 #' Another example: there is no `_mod` suffix available for the `slice_` and `slicev_` methods. \cr \cr
 #'