+++ R/internal_methodchecks.R
@@ -16,13 +16,19 @@
 .methodcheck.i2 <- function(x, abortcall) {
   
   if(is.atomic(x)) {
-    stop("Use the `i_` methods for atomic objects")
+    stop(simpleError(
+      "Use the `i_` methods for atomic objects",
+      call = abortcall
+    ))
   }
   if(is.data.frame(x)) {
-    stop("Use the `ss2_` methods for data.frame-like objects")
+    stop(simpleError(
+      "Use the `ss2_` methods for data.frame-like objects",
+      call = abortcall
+    ))
   }
   if(!is.list(x)) {
-    stop("unsupported object")
+    stop(simpleError("unsupported object", call = abortcall))
   }
   
 }
@@ -31,11 +37,20 @@
 #' @noRd
 .methodcheck.ss <- function(x, abortcall) {
   
+  if(is.null(dim(x))) {
+    stop(simpleError(
+      "cannot use the `ss_`/`ss2_` methods on non-dimensional objects",
+      call = abortcall
+    ))
+  }
   if(is.list(x)) {
-    stop("Use the `ss2_` methods for recursive objects")
+    stop(simpleError(
+      "Use the `ss2_` methods for recursive objects",
+      call = abortcall
+    ))
   }
   if(!is.atomic(x)) {
-    stop("unsupported object")
+    stop(simpleError("unsupported object", call = abortcall))
   }
   
 }
@@ -44,11 +59,20 @@
 #' @noRd
 .methodcheck.ss2 <- function(x, abortcall) {
   
+  if(is.null(dim(x))) {
+    stop(simpleError(
+      "cannot use the `ss_`/`ss2_` methods on non-dimensional objects",
+      call = abortcall
+    ))
+  }
   if(is.atomic(x)) {
-    stop("Use the `ss_` methods for atomic objects")
+    stop(simpleError(
+      "Use the `ss_` methods for atomic objects",
+      call = abortcall
+    ))
   }
   if(!is.list(x)) {
-    stop("unsupported object")
+    stop(simpleError("unsupported object", call = abortcall))
   }
   
 }