+++ inst/examples/generic_rec.R
@@ -20,10 +20,10 @@ lst <- list(
 
 # access recursive subsets ====
 
-i2_rec(lst, c(1,2,2)) # this gives "AA2B"
-i2_rec(lst, c("A", "B", "B")) # this gives "ABB"
-i2_rec(lst, c(2,2,1)) # this gives "BBA"
-i2_rec(lst, c("B", "B", "A")) # this gives "BBA"
+lst_rec(lst, c(1,2,2)) # this gives "AA2B"
+lst_rec(lst, c("A", "B", "B")) # this gives "ABB"
+lst_rec(lst, c(2,2,1)) # this gives "BBA"
+lst_rec(lst, c("B", "B", "A")) # this gives "BBA"
 
 
 #############################################################################
@@ -31,7 +31,7 @@ i2_rec(lst, c("B", "B", "A")) # this gives "BBA"
 # replace recursive subset with R's default in-place semantics ====
 
 # replace "AAB" using R's default in-place semantics:
-i2_recin(
+lst_recin(
   lst, c("A", "A", "B"),
   rp = "THIS IS REPLACED WITH IN-PLACE SEMANTICS"
 )
@@ -43,7 +43,7 @@ print(lst)
 
 # transform recursive subsets with R's default in-place semantics ====
 
-i2_recin(lst, c("C", "A"), tf = \(x)x^2) # transforms lst$C$A
+lst_recin(lst, c("C", "A"), tf = \(x)x^2) # transforms lst$C$A
 
 print(lst)
 
@@ -52,10 +52,10 @@ print(lst)
 
 # add/remove new recursive subsets with R's default in-place semantics ====
 
-i2_recin(lst, c("C", "D"), rp = "NEW VALUE") # adds lst$C$D
+lst_recin(lst, c("C", "D"), rp = "NEW VALUE") # adds lst$C$D
 print(lst)
 
-i2_recin(lst, c("C", "A"), rp = NULL) # removes lst$C$A
+lst_recin(lst, c("C", "A"), rp = NULL) # removes lst$C$A
 print(lst) # notice lst$C$A is GONE
 
 
@@ -68,7 +68,7 @@ x <- list(
  b = data.table::data.table(cola = 11:20, colb = letters[11:20])
 )
 print(x)
-myref <- i2_rec(x, "a")
+myref <- lst_rec(x, "a")
 address(myref) == address(x$a) # they are the same
 ss2_set(myref, vars = "cola", tf = \(x)x^2)
 print(x) # notice x has been changed