+++ inst/tinytest/generic2/source/sourcetest-missingargs.R
@@ -1,11 +1,8 @@
 
-x <- as.list(1:10)
-temp.fun(x)
-
 x <- matrix(as.list(1:20), ncol = 4)
 temp.fun(x)
 
 x <- array(as.list(1:27), c(3,3,3))
 temp.fun(x)
 
-enumerate <- enumerate + 4
+enumerate <- enumerate + 3