+++ inst/tinytest/generic_lst/source/functions4testing.R
@@ -1,0 +1,9 @@
+errorfun <- function(tt) {
+  if(isTRUE(tt)) print(tt)
+  if(isFALSE(tt)) stop(print(tt))
+}
+
+any_empty_indices <- function(...) {
+  lst <- list(...)
+  return(squarebrackets:::.any_empty_indices(lst))
+}