+++ NEWS.md
@@ -1,3 +1,8 @@
 
 # squarebrackets 0.0.0.9000
 * First "real" experimental version of 'squarebrackets' on GitHub.
+
+# squarebrackets 0.0.0.9001
+* Placed the `i2_rec()` and `i2_recin()` in its own, separate set of methods: the `lst_` methods.
+* Adjusted the documentation and tests in accordance with the above change.
+* The `ss_`/`ss2_` methods will now give an error if dimensional sub-set operations are attempted on non-dimensional objects.